requirements.txt
Django
django-cors-headers
django-ninja
django-ninja-jwt[crypto]
pydantic[email]
djangorestframework
channels
daphne
Pillow
celery
dj-database-url
python-dotenv
psycopg2-binary
whitenoise
gunicorn==22.0.0
rav
mssql-django
pyodbc
dj-database-url
<<<<<<< Updated upstream
django-storages[azure]
=======
paypalrestsdk
django-storages
azure-storage-blob
>>>>>>> Stashed changes
